refactor: Extract providers from browser/index in a specific file
